AuthRepositoryからToken取得メソッド等を分離したのでテストコードを修正
