+++ qiita_reader_ios/QiitaReaderTests/Data/Repository/MockTokenRepository.swift
@@ -1,0 +1,26 @@
+//
+//  MockTokenRepository.swift
+//  QiitaReaderTests
+//
+//  Created by Shusuke Ota on 2025/6/3.
+//
+
+import Combine
+import Foundation
+@testable import QiitaReader
+
+class MockTokenRepository: TokenRepositoryProtocol {
+    var accessToken: String?
+
+    func setAccessToken(_ value: String) -> AnyPublisher<Void, Error> {
+        fatalError("Need not to test this method.")
+    }
+
+    func getAccessToken() -> String? {
+        return accessToken
+    }
+
+    func deleteAccessToken() -> AnyPublisher<Void, Error> {
+        fatalError("Need not to test this method.")
+    }
+}
